+++ R/parse_files.R
@@ -788,8 +788,8 @@ PhyDatToMatrix <- function (dataset) {#}, parentheses = c('[', ']'), sep = '') {
   index <- at$index
   allLevels <- as.character(at$allLevels)
   matrix(allLevels[unlist(dataset, recursive = FALSE, use.names = FALSE)],
-           ncol = max(index), byrow = TRUE,
-         dimnames = list(at$names, NULL))[, index, drop = FALSE]
+         ncol = at$nr, byrow = TRUE, dimnames = list(at$names, NULL)
+         )[, index, drop = FALSE]
 }
 
 #' @rdname ReadCharacters