+++ tests/testthat/test-parsers.R
@@ -109,6 +109,20 @@ test_that("Matrix converts to phyDat", {
   expect_equal(mat, PhyDatToMatrix(MatrixToPhyDat(mat)))
 })
 
+test_that("Modified phyDat objects can be converted", {
+  # Obtained by subsetting, i.e. dataset <- biggerDataset[1:4]
+  dataset <- structure(list(a = c(1L, 1L, 1L, 1L), c = c(2L, 1L, 1L, 2L),
+                            d = c(2L, 1L, 2L, 1L), e = c(2L, 2L, 2L, 2L)),
+                       weight = c(3L, 3L, 0L, 0L), nr = 4L, nc = 2L,
+                       index = c(2L, 1L, 1L, 1L, 2L, 2L),
+                       .Label = c("0", "1"), allLevels = c("0", "1"),
+                       type = "USER",
+                       contrast = structure(c(1, 0, 0, 1), .Dim = c(2L, 2L),
+                                            .Dimnames = list(NULL, c("0", "1"))),
+                       class = "phyDat")
+  expect_equal(c(4, 6), dim(PhyDatToMatrix(dataset)))
+})
+
 test_that("MatrixToPhyDat() warns when characters blank", {
   # May occur when loading an excel file with empty cells
   mat <- matrix(c(1,0,1,0,1,0,1,0,0,'','','',0,1,0,1,2,2,2,2,2,2,2,'?'),